+++ test/forwarding.test.js
@@ -1,77 +1,78 @@
 var GsTb = require(process.env.GSTB_COV ? '../gstb-cov.js' : '../')
   , GS = require('groupstage');
 
-exports.forwardingSixteen = function (t) {
+exports.sixteenFourForwarding = function (t) {
   var trn = new GsTb(16, { groupSize: 4, limit: 4 });
-  t.ok(!trn.stageComplete(), 'need to play first round');
+  t.ok(!trn.isStageComplete(), 'need to play first round');
 
-  var ensureMiddleBoundaries = function (round) {
+  var ensureMiddleBoundaries = function () {
     t.ok(!trn.isDone(), "whole tournament not done");
-    t.ok(round.isDone(), "round complete");
-    t.ok(trn.stageComplete(), 'can start next stage now');
-    t.ok(trn.createNextStage(), "could create next stage");
-    t.ok(!trn.stageComplete(), 'need to play second round');
-    t.ok(!trn.currentRound().isDone(), "next round incomplete");
+    t.ok(trn.isStageComplete(), 'stage done');
+    t.ok(trn.propagate(), "could create next stage");
+    t.ok(!trn.isStageComplete(), 'need to play second round');
   };
 
-  var gs = trn.currentRound();
+  var msGs = trn.active();
   var expR1 = GS(16, { groupSize: 4 }).matches;
-  t.deepEqual(gs.matches, expR1, "Stage 1 === orig GS");
+  t.deepEqual(msGs, expR1, "Stage 1 === orig GS");
 
   // score s.t. tiebreakers fully necessary
-  gs.matches.forEach(function (m) {
-    gs.score(m.id, [1, 1]);
+  msGs.forEach(function (m) {
+    trn.score(m.id, [1, 1]);
   });
 
-  ensureMiddleBoundaries(gs);
-  t.ok(trn.isTieBreakerRound(), 'we should be tied now');
+  ensureMiddleBoundaries();
+  t.ok(trn.isTieBreakerRound(), 'we should be tied now 1');
 
 
-  var tb = trn.currentRound();
+  var msTb = trn.active();
   var expR2 = expR1.slice(); // no resolution at all in gs so tb is equivalent
-  t.deepEqual(tb.matches, expR2, "Stage 2 === orig GS in TB form");
+  t.deepEqual(msTb, expR2, "Stage 2 === orig GS in TB form");
 
-  tb.matches.forEach(function (m) {
+  msTb.forEach(function (m) {
     if (m.id.s === 1) {
       // keep tieing group 1
-      tb.score(m.id, [1,1]);
+      trn.score(m.id, [1,1]);
     }
     else {
-      tb.score(m.id, m.p[0] < m.p[1] ? [1,0] : [0,1]);
+      trn.score(m.id, m.p[0] < m.p[1] ? [1,0] : [0,1]);
     }
   });
 
-  ensureMiddleBoundaries(tb);
-  t.ok(trn.isTieBreakerRound(), 'we should still be tied');
+  ensureMiddleBoundaries();
+  t.ok(trn.isTieBreakerRound(), 'we should still be tied 2');
 
-  var tb2 = trn.currentRound();
+  var msTb2 = trn.active()
   // know this is sufficient to verify it's a TB because 1st placers not present
   var expR3 = expR2.slice().filter(function (m) {
     return m.id.s === 1;
   });
-  t.deepEqual(tb2.matches, expR3, "Stage 3 === Group 1 TB");
+  t.deepEqual(msTb2, expR3, "Stage 3 === Group 1 TB");
   t.equal(expR3.length, GS(4).matches.length, "length equivalent to a 4p GS");
-  t.equal(tb2.players().length, 4, "4 players left");
+  // TODO: bad use of this.current
+  t.equal(trn.current.players().length, 4, "4 players left");
 
-  tb2.matches.forEach(function (m) {
+  msTb2.forEach(function (m) {
     // reduce num players for next
-    tb2.score(m.id, (m.id.m === 1) ? [1,1] : [1,0]);
+    trn.score(m.id, (m.id.m === 1) ? [1,1] : [1,0]);
   });
 
-  ensureMiddleBoundaries(tb2);
+  ensureMiddleBoundaries();
+  t.ok(trn.isTieBreakerRound(), 'we should still be tied 3');
 
-  var tb3 = trn.currentRound();
-  t.deepEqual(tb3.players(), [5,12,16], '2nd placers in grp 1');
+  var msTb3 = trn.active();
+  // TODO: use of this.current should not be needed - although this is helpful:
+  t.deepEqual(trn.current.players(), [5,12,16], '2nd placers in grp 1');
 
-  tb3.matches.forEach(function (m) {
-    tb3.score(m.id, m.p[0] < m.p[1] ? [1,0] : [0,1]); // resolve rest
+  msTb3.forEach(function (m) {
+    trn.score(m.id, m.p[0] < m.p[1] ? [1,0] : [0,1]); // resolve rest
   });
 
   // ensure everthing done - no middle boundry check this time
-  t.ok(tb3.isDone(), "round complete");
-  t.ok(trn.stageComplete(), 'can start next stage now');
-  t.ok(!trn.createNextStage(), "could not create any more stages now");
-  t.ok(trn.isDone(), 'the whole tournament is done');
+  t.ok(trn.isStageComplete(), 'final stage complete');
+  t.ok(trn.isDone(), "and tourney complete");
+  
+  t.ok(!trn.propagate(), "thus propagate fails");
 
   t.done();
 };